refactor: move formating functions in separate module
