+++ src/main.rs
@@ -9,12 +9,14 @@ use std::time::{Duration, SystemTime};
 use chrono::prelude::*;
 use clap::Parser;
 use reqwest::blocking::Client;
-use serde_json::{json, Map, Value};
+use serde_json::{json, Value};
 
 use crate::cli::Args;
+use crate::format::{format_ampm_time, format_chances, format_indicator, format_temp, format_time};
 use crate::lang::Lang;
 
 mod cli;
+mod format;
 mod lang;
 
 const WEATHER_CODES: &[(i32, &str)] = &[
@@ -285,99 +287,4 @@ fn main() {
 
     let json_data = json!(data);
     println!("{}", json_data);
-}
-
-fn format_time(time: &str, ampm: bool) -> String {
-    let hour = time.replace("00", "").parse::<i32>().unwrap();
-
-    if ampm {
-        let am_or_pm = if hour >= 12 { "pm" } else { "am" };
-        let hour12 = if hour == 0 || hour == 12 {
-            12
-        } else {
-            hour % 12
-        };
-        format!("{: <4}", format!("{}{}", hour12, am_or_pm))
-    } else {
-        format!("{:02}", hour)
-    }
-}
-
-fn format_temp(temp: &str) -> String {
-    format!("{: >3}°", temp)
-}
-
-fn format_chances(hour: &serde_json::Value, lang: &Lang) -> String {
-    let chances: HashMap<&str, String> = [
-        ("chanceoffog", lang.fog()),
-        ("chanceoffrost", lang.frost()),
-        ("chanceofovercast", lang.overcast()),
-        ("chanceofrain", lang.rain()),
-        ("chanceofsnow", lang.snow()),
-        ("chanceofsunshine", lang.sunshine()),
-        ("chanceofthunder", lang.thunder()),
-        ("chanceofwindy", lang.wind()),
-    ]
-    .iter()
-    .cloned()
-    .collect();
-
-    let mut conditions = vec![];
-    for (event, name) in chances.iter() {
-        if let Some(chance) = hour[event].as_str() {
-            if let Ok(chance_value) = chance.parse::<u32>() {
-                if chance_value > 0 {
-                    conditions.push((name, chance_value));
-                }
-            }
-        }
-    }
-    conditions.sort_by_key(|&(_, chance_value)| std::cmp::Reverse(chance_value));
-    conditions
-        .iter()
-        .map(|&(name, chance_value)| format!("{} {}%", name, chance_value))
-        .collect::<Vec<_>>()
-        .join(", ")
-}
-
-fn format_ampm_time(day: &serde_json::Value, key: &str, ampm: bool) -> String {
-    if ampm {
-        day["astronomy"][0][key].as_str().unwrap().to_string()
-    } else {
-        NaiveTime::parse_from_str(day["astronomy"][0][key].as_str().unwrap(), "%I:%M %p")
-            .unwrap()
-            .format("%H:%M")
-            .to_string()
-    }
-}
-fn format_indicator(weather_conditions: &Value, expression: String, weather_icon: &&str) -> String {
-    if !weather_conditions.is_object() {
-        return String::new();
-    }
-    let default_map = Map::new();
-    let weather_conditions_map = weather_conditions.as_object().unwrap_or(&default_map);
-    let mut formatted_indicator = expression.to_string();
-    weather_conditions_map
-        .iter()
-        .map(|condition| ("{".to_owned() + condition.0 + "}", condition.1))
-        .for_each(|condition| {
-            if formatted_indicator.contains(condition.0.as_str()) {
-                let condition_value = if condition.1.is_array() {
-                    condition.1.as_array().and_then(|vec| {
-                        vec[0]
-                            .as_object()
-                            .and_then(|value_map| value_map["value"].as_str())
-                    })
-                } else {
-                    condition.1.as_str()
-                }
-                .unwrap_or("");
-                formatted_indicator =
-                    formatted_indicator.replace(condition.0.as_str(), condition_value)
-            }
-        });
-    if formatted_indicator.contains(ICON_PLACEHOLDER) {
-        formatted_indicator = formatted_indicator.replace(ICON_PLACEHOLDER, weather_icon)
-    }
-    formatted_indicator
 }
